+++ tests/libtest/lib1571.c
@@ -50,7 +50,7 @@ CURLcode test(char *URL)
     test_setopt(curl, CURLOPT_POSTFIELDS, "moo");
   }
   if(testno == 1581) {
-    test_setopt(curl, CURLOPT_POSTREDIR, CURL_REDIR_POST_301);
+    test_setopt(curl, CURLOPT_POSTREDIR, (long)CURL_REDIR_POST_301);
   }
 
   test_setopt(curl, CURLOPT_CUSTOMREQUEST, "IGLOO");
